sycl : refactored helper headers into multiple files
